+++ src/pages/nemesis_pages/nemesis_page_ids.py
@@ -1,0 +1,11 @@
+
+dropdown_user_selection = 'dropdown-user-selection-nemesis'
+dropdown_against_selection = 'dropdown-against-selection-nemesis'
+
+dropdown_match_type_selection = 'dropdown-match-type-selection-nemesis'
+
+filtered_nemesis_df = 'filtered-nemesis-df'
+filtered_against_df = 'filtered-nemesis--against-df'
+
+nemesis = 'nemesis'
+opponent_stats = 'nemesis-opponent-stats'